First cut at sil shading of neutrals - not working yet

+++ src/cmd4.c
@@ -660,7 +660,7 @@ void do_cmd_load_screen(void)
 {
   int i, y, x;
   
-  byte a = 0;
+  int a = 0;
   wchar_t c = L' ';
   
   bool okay = TRUE;
@@ -750,7 +750,7 @@ void do_cmd_save_screen_text(void)
 {
   int y, x;
   
-  byte a = 0;
+  int a = 0;
   wchar_t c = L' ';
   
   ang_file *fff;
